// src/app/interfaces/user.interface.ts
export interface User {
    userId:        number;
    name:          string;
    email:         string;
    address:       string;
    phone:         string;
    created:       string;
    updated:       string;
    accountNumber: string;
    accountType:   string;
    bank:          string;
    balance:       number;
    accountId:     number;
    token:         string;
}